partial lifestream IPC integration, nameplate alteration implementation. Injected native game handles for  gagged icon displays.

+++ ProjectGagSpeak/PlayerClient/Configs/ServerConfigs.cs
@@ -167,7 +167,7 @@ public record Authentication
 {
     public ulong CharacterPlayerContentId { get; set; } = 0;
     public string CharacterName { get; set; } = string.Empty;
-    public uint WorldId { get; set; } = 0;
+    public ushort WorldId { get; set; } = 0;
     public bool IsPrimary { get; set; } = false;
     public SecretKey SecretKey { get; set; } = new();
 }
